ECO-468: fix bug of requesting twice when clicking pagination.
