+++ explorer/ui/src/components/BlockList.tsx
@@ -25,10 +25,7 @@ export interface Props extends RouteComponentProps<{}> {
 class _BlockList extends React.Component<Props, {}> {
   constructor(props: Props) {
     super(props);
-    let maxRank = parseInt(props.maxRank || '') || 0;
-    let depth = parseInt(props.depth || '') || 10;
-    this.props.dag.updateMaxRankAndDepth(maxRank, depth);
-    this.props.dag.refreshBlockDagAndSetupSubscriber();
+    this.props.dag.refreshWithDepthAndMaxRank(props.maxRank, props.depth);
   }
 
   async refresh() {
@@ -38,6 +35,13 @@ class _BlockList extends React.Component<Props, {}> {
   componentWillUnmount() {
     // release websocket if necessary
     this.props.dag.toggleableSubscriber.unsubscribeAndFree();
+  }
+
+  componentWillReceiveProps(nextProps: Props) {
+    if (this.props.depth === nextProps.depth && this.props.maxRank === nextProps.maxRank) {
+      return;
+    }
+    this.props.dag.refreshWithDepthAndMaxRank(nextProps.maxRank, nextProps.depth);
   }
 
   render() {
