+++ explorer/ui/src/components/Explorer.tsx
@@ -16,31 +16,21 @@ import { BondedValidatorsTable } from './BondedValidatorsTable';
 import { ToggleButton } from './ToggleButton';
 import { BlockType, BlockRole, FinalityIcon } from './BlockDetails';
 
-const DEFAULT_DEPTH = 100;
 
 /** Show the tips of the DAG. */
 @observer
 class _Explorer extends React.Component<Props, {}> {
   constructor(props: Props) {
     super(props);
-    this.refreshWithDepthAndMaxRank(props.maxRank, props.depth);
+    this.props.dag.refreshWithDepthAndMaxRank(props.maxRank, props.depth);
   }
 
-  refreshWithDepthAndMaxRank(
-    maxRankStr: string | null,
-    depthStr: string | null
-  ) {
-    let maxRank = parseInt(maxRankStr || '') || 0;
-    let depth = parseInt(depthStr || '') || DEFAULT_DEPTH;
-    this.props.dag.updateMaxRankAndDepth(maxRank, depth);
-    this.props.dag.refreshBlockDagAndSetupSubscriber();
-  }
 
   componentWillReceiveProps(nextProps: Props) {
     if (this.props.depth === nextProps.depth && this.props.maxRank === nextProps.maxRank) {
       return;
     }
-    this.refreshWithDepthAndMaxRank(nextProps.maxRank, nextProps.depth);
+    this.props.dag.refreshWithDepthAndMaxRank(nextProps.maxRank, nextProps.depth);
   }
 
   async refresh() {
